Refactors and integrates OpenAI API endpoints (#72)
* Restructures and refactors API endpoints

Refactors the API structure by renaming and moving files for better organization.

Updates the API configuration to a dedicated config.json file and modifies the application name.

Adjusts API routes to provide more descriptive information and interact with external services like Flickr.

* Refactors and integrates OpenAI API endpoints

Reorganizes the API structure to incorporate OpenAI functionalities.

- Renames and moves files related to CV analysis to reflect the integration of OpenAI.
- Updates the main API endpoint to include a reference to the new OpenAI endpoint.
- Removes unused or deprecated endpoints.
