+++ app/api/gl-api/openai/cv/scrap.md
